Fix bug that building with TI_EXPORT_CORE:BOOL=ON failed
