+++ taichi/codegen/codegen_llvm.cpp
@@ -1,4 +1,5 @@
 #ifdef TI_WITH_LLVM
+#include "taichi/analysis/offline_cache_util.h"
 #include "taichi/codegen/codegen_llvm.h"
 #include "taichi/llvm/llvm_offline_cache.h"
 #include "taichi/ir/statements.h"
