+++ taichi/ir/expression_printer.h
@@ -4,7 +4,7 @@
 #include "taichi/ir/expression.h"
 #include "taichi/ir/frontend_ir.h"
 #include "taichi/program/program.h"
-#include "taichi/llvm/llvm_offline_cache.h"
+#include "taichi/analysis/offline_cache_util.h"
 
 namespace taichi {
 namespace lang {
